[TASK] Move pageUrls file and fix userscript URL permissions
