+++ webpackConfig/webextension.assets.js
@@ -4,7 +4,7 @@ const fs = require('fs');
 const mkdirp = require('mkdirp');
 const download = require('download-file');
 const playerUrls = require('../src/pages/playerUrls');
-const pageUrls = require('../src/pages/pageUrls');
+const pageUrls = require('./utils/pageUrls');
 const packageJson = require('../package.json');
 const resourcesJson = require('./resources');
 const i18n = require('./utils/i18n');
